feat(interpreter): add support for BOUNCE opcode
This is Goofy's equivalent of JUMP where we jump if a certain condition is made to a defined label. To make this change easier to use the outer loop has been replaced with a while loop so we can change the index as we see fit. Some additional helper methods have been added for checking the stack.